webapi/tct-mediaqueries-css3-tests/mediaqueries/CSS3MQ_width_screen_positive.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Wang,Ling <lingx.z.wang@intel.com>
        Cui,Jieqiong <jieqiongx.cui@intel.com>

-->

<html>
  <head>
    <title>CSS3MediaQueries Test: CSS3MQ_width_screen_positive</title>
    <link rel="author" title="Intel" href="http://www.intel.com/" />
    <link rel="help" href="http://www.w3.org/TR/2012/REC-css3-mediaqueries-20120619/#width" />
    <meta name="flags" content="" />
    <meta name="assert" content="Set 'width: 800px' on CSS3 Media Queries width attr, When the browser width is 800px the text will displays in orange" />
    <style>
      @media  screen and (width: 800px) {
        body{color:#f60;}
      }
    </style>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <div id="media">Set "width: 800px" on CSS3 Media Queries width attr, When the broswer width is 800px the text will displays in orange</div>
    <script>
      test(function () {
        var color = getComputedStyle(document.getElementById("media"))["color"];
        if(document.body.scrollWidth == 800) {
          assert_equals(color, "rgb(255, 102, 0)", "The text color");
        } else{
          assert_equals(color, "rgb(0, 0, 0)", "The text color");
        }
      })
    </script>
  </body>
</html>
